requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
fake-useragent==0.1.11
idna==3.3
loguru==0.6.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.0
rsa==4.8
toml==0.10.2
urllib3==1.26.9
win32-setctime==1.1.0
